⚡ client app compiles and runs after restructure
