+++ Decsys/ClientApp/.env
@@ -1,0 +1,1 @@
+SKIP_PREFLIGHT_CHECK=true
