+++ Decsys/ClientApp/src/app/screens/admin/SurveysScreen/SurveysScreen.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { useNavigation } from "react-navi";
-import { useNaviReducer } from "utils/hooks";
+import { useNaviReducer } from "hooks/reducers";
 import PureSurveysScreen from "./PureSurveysScreen";
 import SurveyCardContext from "components/SurveyCard/Context";
 import reducer, * as ducks from "./ducks";
